ready to deploy

+++ app.js
@@ -7,13 +7,21 @@ const app = express();
 const port = process.env.NODE_ENV === "production" ? 5000 : 3000;
 const host = process.env.NODE_ENV === "production" ? "0.0.0.0" : "localhost";
 
+function sanitize(str) {
+  return str.trim().replace(/\s+/g, "");
+}
+
 app.set("view engine", "ejs");
 app.use(layouts);
 app.use(express.json());
 app.use(express.urlencoded({ extended: true }));
 
+function sanitize(str) {
+  return str.trim().replace(/\s+/g, " ");
+}
+
 let sql;
-app.get("/product", (req, res) => {
+app.get("/", (req, res) => {
   sql = `SELECT 
   product.id AS id,
   product.name AS name,
@@ -41,35 +49,50 @@ GROUP BY name`;
     });
   });
 });
-app.get("/product/add", (req, res) => {
+app.get("/add", (req, res) => {
   res.status(200).render("add", { title: "Insert", layout: "layouts/main" });
 });
-app.post("/product/add", (req, res) => {
+app.post("/add", (req, res) => {
   const { name, price, description, category } = req.body;
-  sql = `INSERT INTO product VALUES (0, '${name}', '${price}', '${description}')`;
+  if (name === "" || price === "" || description === "" || !category) {
+    return res.status(400).redirect("add");
+  }
+
+  sql = `INSERT INTO product VALUES (0, '${sanitize(name)}', '${sanitize(
+    price
+  )}', '${sanitize(description)}')`;
   db.query(sql, (err, result) => {
     if (err) {
-      res.status(400).redirect("/product");
+      res.status(400).redirect("/");
       throw err;
     }
     const insertedId = result.insertId;
-    const promises = category.map((selectedCategory) => {
-      sql = `INSERT INTO product_category VALUES (${insertedId}, ${selectedCategory})`;
-      return new Promise((resolve, reject) => {
-        db.query(sql, (err) => {
-          if (err) reject(err);
-          resolve();
+    if (!Array.isArray(category)) {
+      sql = `INSERT INTO product_category VALUES (${insertedId}, ${category})`;
+      db.query(sql, (err) => {
+        if (err) throw err;
+        return res.redirect("/");
+      });
+    } else {
+      const promises = category.map((selectedCategory) => {
+        sql = `INSERT INTO product_category VALUES (${insertedId}, ${selectedCategory})`;
+        return new Promise((resolve, reject) => {
+          db.query(sql, (err) => {
+            if (err) reject(err);
+            resolve();
+          });
         });
       });
-    });
-    Promise.all(promises)
-      .then(res.redirect("/product"))
-      .catch((err) => {
-        throw err;
-      });
+      Promise.all(promises)
+        .then(res.redirect("/"))
+        .catch((err) => {
+          throw err;
+        });
+    }
   });
 });
-app.get("/product/delete/:id", (req, res) => {
+
+app.get("/delete/:id", (req, res) => {
   const id = req.params.id;
   sql = `DELETE FROM product_category WHERE product_category.id_product = ${id}`;
   db.query(sql, (err) => {
@@ -77,16 +100,16 @@ app.get("/product/delete/:id", (req, res) => {
     sql = `DELETE FROM product WHERE id = ${id}`;
     db.query(sql, (err) => {
       if (err) throw err;
-      res.redirect("/product");
+      res.redirect("/");
     });
   });
 });
-app.get("/product/update/:id", (req, res) => {
+app.get("/update/:id", (req, res) => {
   const id = req.params.id;
   sql = `SELECT product.id AS id, product.name AS name, product.price AS price, product.description AS description, group_concat(pc.id_category separator ", ") AS productTags, GROUP_CONCAT(category.name SEPARATOR ", ") AS category FROM product JOIN product_category AS pc ON (id = pc.id_product) JOIN category ON (pc.id_category = category.id) WHERE product.id = ${id} GROUP BY name`;
   db.query(sql, (err, result) => {
     if (err) {
-      res.redirect("/product");
+      res.redirect("/");
       throw err;
     }
     res.render("update", {
@@ -96,24 +119,46 @@ app.get("/product/update/:id", (req, res) => {
     });
   });
 });
-app.post("/product/update", (req, res) => {
+app.post("/update", (req, res) => {
   const { id, name, price, description, category } = req.body;
+  if (name === "" || price === "" || description === "" || !category) {
+    return res.status(400).redirect(`/update/${id}`);
+  }
   sql = `DELETE FROM product_category WHERE id_product = ${id}`;
   db.query(sql, (err) => {
     if (err) throw err;
-  });
-  sql = `UPDATE product SET name = ${name}', price = '${price}', description = '${description}')`;
-  db.query(sql, (err, result) => {
-    let promises = category.map((tag) => {
-      sql = `INSERT INTO product_category values (${id}, ${tag})`;
-      return new Promise((resolve, reject) => {
-        db.query(sql, (err, result) => {
-          if (err) reject(err);
-          resolve();
+    sql = `UPDATE product SET name = '${sanitize(name)}', price = '${sanitize(
+      price
+    )}', description = '${sanitize(description)}' WHERE id = '${id}'`;
+    db.query(sql, (err, result) => {
+      if (err) {
+        res.status(400).redirect("/");
+        throw err;
+      }
+      const insertedId = result.insertId;
+      if (!Array.isArray(category)) {
+        sql = `INSERT INTO product_category VALUES (${id}, ${category})`;
+        db.query(sql, (err) => {
+          if (err) throw err;
+          return res.redirect("/");
         });
-      });
+      } else {
+        const promises = category.map((selectedCategory) => {
+          sql = `INSERT INTO product_category VALUES (${id}, ${selectedCategory})`;
+          return new Promise((resolve, reject) => {
+            db.query(sql, (err) => {
+              if (err) reject(err);
+              resolve();
+            });
+          });
+        });
+        Promise.all(promises)
+          .then(res.redirect("/"))
+          .catch((err) => {
+            throw err;
+          });
+      }
     });
-    Promise.all(promises).then(() => res.redirect("/product"));
   });
 });
 app.listen(port, host, () => console.log("app runnning"));
